split ErrorBoundary into two components and wrapped the entire main page in it
