+++ src/components/ErrorBoundary/index.tsx
@@ -1,0 +1,1 @@
+export { AppErrorBoundary } from "./ErrorBonduary";
