+++ src/components/MainPage/MainPage.tsx
@@ -1,5 +1,5 @@
-import { AppErrorBoundary } from "@/ErrorBonduary";
-import { ErrorTrigger } from "@components/ErrorTrigger";
+import { AppErrorBoundary } from "@components/ErrorBoundary";
+import { ErrorTrigger } from "@components/ErrorBoundary/ErrorTrigger";
 import { Header } from "@components/Header";
 import TopicRenderer from "@components/TopicRenderer";
 
@@ -7,8 +7,8 @@ export const MainPage = () => {
   return (
     <div>
       <Header />
-      <TopicRenderer />
       <AppErrorBoundary>
+      <TopicRenderer />
         <ErrorTrigger/>
       </AppErrorBoundary>
     </div>
